+++ backend/Core/Core.BLL/Services/CurrencyTransferService.cs
@@ -7,23 +7,32 @@ using Microsoft.AspNetCore.Http.HttpResults;
 
 namespace Core.BLL.Services;
 
-public class CurrencyTransferService {
-    public async Task<int> TransferCurrency(CurrencyType fromCurrency, CurrencyType toCurrency, int amount) {
+public class CurrencyTransferService
+{
+    public async Task<long> TransferCurrency(
+        CurrencyType fromCurrency,
+        CurrencyType toCurrency,
+        long amount
+    )
+    {
         if (fromCurrency == toCurrency)
             return amount;
         using HttpClient client = new HttpClient();
-       
-            var stringResponse = await client.GetStringAsync("https://v6.exchangerate-api.com/v6/fec28a804fd42646bb06b88f/latest/USD");
-            CurrencyApiDto? response = JsonSerializer.Deserialize<CurrencyApiDto>(stringResponse);
-            if (response is not { result: "success" }) 
-                throw new InvalidOperationException();
-            
-            var fromRate = Math.Round(GetConversionRate(fromCurrency, response), 2);
-            var toRate = Math.Round(GetConversionRate(toCurrency, response), 2);
-            
-            double convertedAmount = ((amount/ 100.0) / fromRate) * toRate;
-            return (int)Math.Round(convertedAmount * 100, 2);
+
+        var stringResponse = await client.GetStringAsync(
+            "https://v6.exchangerate-api.com/v6/fec28a804fd42646bb06b88f/latest/USD"
+        );
+        CurrencyApiDto? response = JsonSerializer.Deserialize<CurrencyApiDto>(stringResponse);
+        if (response is not { result: "success" })
+            throw new InvalidOperationException();
+
+        var fromRate = Math.Round(GetConversionRate(fromCurrency, response), 2);
+        var toRate = Math.Round(GetConversionRate(toCurrency, response), 2);
+
+        double convertedAmount = ((amount / 100.0) / fromRate) * toRate;
+        return (long)Math.Round(convertedAmount * 100, 2);
     }
+
     static double GetConversionRate(CurrencyType currencyType, CurrencyApiDto currencyData)
     {
         switch (currencyType)
