+++ backend/Loan/Loan.BLL/DataTransferObjects/LoanChargeDto.cs
@@ -3,14 +3,18 @@ using Common.Enum;
 
 namespace Loan.BLL.DataTransferObjects;
 
-public class LoanChargeDto {
+public class LoanChargeDto
+{
     [Required]
-    public int Amount { get; set; }
+    public long Amount { get; set; }
+
     [Required]
     public Guid LoanId { get; set; }
+
     [Required]
     public Guid AccountId { get; set; }
     public Guid UserId { get; set; }
+
     [Required]
     public CurrencyType CurrencyType { get; set; }
 }
